refactor: dto 네이밍 변경
리뷰 반영
