+++ pumping/src/main/kotlin/com/dpm/pumping/auth/application/AuthService.kt
@@ -2,9 +2,9 @@ package com.dpm.pumping.auth.application
 
 import com.dpm.pumping.auth.domain.LoginPlatform
 import com.dpm.pumping.auth.domain.LoginType
-import com.dpm.pumping.auth.dtos.AccessTokenResponse
-import com.dpm.pumping.auth.oauth2.dtos.OAuth2LoginResponse
-import com.dpm.pumping.auth.oauth2.dtos.SignUpRequest
+import com.dpm.pumping.auth.dto.AccessTokenResponse
+import com.dpm.pumping.auth.oauth2.dto.OAuth2LoginResponse
+import com.dpm.pumping.auth.oauth2.dto.SignUpRequest
 import com.dpm.pumping.user.domain.User
 import com.dpm.pumping.user.domain.UserRepository
 import org.springframework.stereotype.Service
@@ -26,10 +26,10 @@ class AuthService(
             }
         val token = jwtTokenProvider.generateAccessToken(user.uid)
         return OAuth2LoginResponse(
-            token.accessToken,
-            token.expiredTime,
-            user.platform.loginType,
-            user.platform.oauth2Id
+            accessToken = token.accessToken,
+            expiredTime = token.expiredTime,
+            loginType = user.platform.loginType,
+            oauth2Id = user.platform.oauth2Id
         )
     }
 
@@ -40,11 +40,11 @@ class AuthService(
             "인증되지 않은 사용자입니다. oauth 로그인을 다시하세요"
         )
         user.register(
-            request.name,
-            request.gender,
-            request.height,
-            request.weight,
-            platform
+            name = request.name,
+            gender = request.gender,
+            height = request.height,
+            weight = request.weight,
+            platform = platform
         )
         userRepository.save(user)
         return jwtTokenProvider.generateAccessToken(user.uid)
