+++ pumping/src/main/kotlin/com/dpm/pumping/auth/presentation/AuthController.kt
@@ -1,12 +1,12 @@
 package com.dpm.pumping.auth.presentation
 
 import com.dpm.pumping.auth.application.AuthService
-import com.dpm.pumping.auth.dtos.*
+import com.dpm.pumping.auth.dto.*
 import com.dpm.pumping.auth.oauth2.OAuth2AppleClient
-import com.dpm.pumping.auth.oauth2.dtos.AppleLoginRequest
-import com.dpm.pumping.auth.oauth2.dtos.AppleLoginUrlResponse
-import com.dpm.pumping.auth.oauth2.dtos.OAuth2LoginResponse
-import com.dpm.pumping.auth.oauth2.dtos.SignUpRequest
+import com.dpm.pumping.auth.oauth2.dto.AppleLoginRequest
+import com.dpm.pumping.auth.oauth2.dto.AppleLoginUrlResponse
+import com.dpm.pumping.auth.oauth2.dto.OAuth2LoginResponse
+import com.dpm.pumping.auth.oauth2.dto.SignUpRequest
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
