+++ pumping/src/test/kotlin/com/dpm/pumping/auth/presentation/AuthControllerTest.kt
@@ -3,13 +3,13 @@ package com.dpm.pumping.auth.presentation
 import com.dpm.pumping.auth.application.AuthService
 import com.dpm.pumping.auth.application.JwtTokenProvider
 import com.dpm.pumping.auth.domain.LoginType
-import com.dpm.pumping.auth.dtos.AccessTokenResponse
+import com.dpm.pumping.auth.dto.AccessTokenResponse
 import com.dpm.pumping.auth.oauth2.OAuth2AppleClaimsValidator
 import com.dpm.pumping.auth.oauth2.OAuth2AppleClient
-import com.dpm.pumping.auth.oauth2.dtos.AppleLoginRequest
-import com.dpm.pumping.auth.oauth2.dtos.AppleLoginUrlResponse
-import com.dpm.pumping.auth.oauth2.dtos.OAuth2LoginResponse
-import com.dpm.pumping.auth.oauth2.dtos.SignUpRequest
+import com.dpm.pumping.auth.oauth2.dto.AppleLoginRequest
+import com.dpm.pumping.auth.oauth2.dto.AppleLoginUrlResponse
+import com.dpm.pumping.auth.oauth2.dto.OAuth2LoginResponse
+import com.dpm.pumping.auth.oauth2.dto.SignUpRequest
 import com.dpm.pumping.user.domain.Gender
 import com.fasterxml.jackson.databind.ObjectMapper
 import org.junit.jupiter.api.Assertions.*
